# Conflicts:
#	beacon_node/execution_layer/src/lib.rs
#	beacon_node/http_api/src/publish_blocks.rs
#	lcli/src/new_testnet.rs

+++ beacon_node/lighthouse_network/src/service/gossipsub_scoring_parameters.rs
@@ -270,11 +270,11 @@ impl<TSpec: EthSpec> PeerScoreSettings<TSpec> {
 
         let modulo_smaller = max(
             1,
-            smaller_committee_size / self.target_aggregators_per_committee as usize,
+            smaller_committee_size / self.target_aggregators_per_committee,
         );
         let modulo_larger = max(
             1,
-            (smaller_committee_size + 1) / self.target_aggregators_per_committee as usize,
+            (smaller_committee_size + 1) / self.target_aggregators_per_committee,
         );
 
         Ok((
